call_log: extract find_participant to its own module
Why:

* Will not be used by cel_interpretor anymore
