+++ wazo_call_logd/cel_interpretor.py
@@ -6,77 +6,15 @@ import re
 import logging
 
 from xivo.asterisk.line_identity import identity_from_channel
-from xivo.asterisk.protocol_interface import protocol_interface_from_channel
-from xivo.asterisk.protocol_interface import InvalidChannelError
 
 from .database.cel_event_type import CELEventType
 from .database.models import CallLogParticipant, Recording
+from .participant import find_participant
 
 
 logger = logging.getLogger(__name__)
 
 EXTRA_USER_FWD_REGEX = r'^.*NUM: *(.*?) *, *CONTEXT: *(.*?) *, *NAME: *(.*?) *(?:,|"})'
-
-
-def find_participant(confd, channame):
-    # TODO PJSIP clean after migration
-    channame = channame.replace('PJSIP', 'SIP')
-
-    try:
-        protocol, line_name = protocol_interface_from_channel(channame)
-    except InvalidChannelError:
-        return None
-
-    if protocol == 'Local':
-        logger.debug('Ignoring participant %s', channame)
-        return None
-
-    logger.debug(
-        'Looking up participant with protocol %s and line name "%s"',
-        protocol,
-        line_name,
-    )
-    lines = confd.lines.list(name=line_name, recurse=True)['items']
-    if not lines:
-        return
-
-    line = lines[0]
-    logger.debug('Found participant line id %s', line['id'])
-    users = line['users']
-    if not users:
-        return
-
-    user = confd.users.get(users[0]['uuid'])
-    tags = (
-        [tag.strip() for tag in user['userfield'].split(',')]
-        if user['userfield']
-        else []
-    )
-    logger.debug(
-        'Found participant user uuid %s tenant uuid %s',
-        user['uuid'],
-        user['tenant_uuid'],
-    )
-
-    extensions = line['extensions']
-    if extensions:
-        main_extension = extensions[0]
-
-        logger.debug(
-            'Found main internal extension %s@%s',
-            main_extension['exten'],
-            main_extension['context'],
-        )
-    else:
-        main_extension = None
-
-    return {
-        'uuid': user['uuid'],
-        'tenant_uuid': user['tenant_uuid'],
-        'line_id': line['id'],
-        'tags': tags,
-        'main_extension': main_extension,
-    }
 
 
 def extract_mixmonitor_extra(extra):
