+++ wazo_call_logd/tests/test_cel_interpretor.py
@@ -17,7 +17,6 @@ from ..cel_interpretor import (
     AbstractCELInterpretor,
     CallerCELInterpretor,
     DispatchCELInterpretor,
-    find_participant,
     extract_mixmonitor_extra,
     is_valid_mixmonitor_start_extra,
     is_valid_mixmonitor_stop_extra,
@@ -34,55 +33,6 @@ def confd_mock(lines=None):
         lines[0]['users'][0] if lines and lines[0].get('users') else None
     )
     return confd
-
-
-class TestFindParticipant(TestCase):
-    def test_find_participants_when_channame_is_not_parsable(self):
-        confd = confd_mock()
-        channame = 'something'
-
-        result = find_participant(confd, channame)
-
-        assert_that(result, none())
-
-    def test_find_participants_when_no_lines(self):
-        confd = confd_mock()
-        channame = 'sip/something-suffix'
-
-        result = find_participant(confd, channame)
-
-        assert_that(result, none())
-
-    def test_find_participants_when_line_has_no_user(self):
-        lines = [{'id': 12, 'users': []}]
-        confd = confd_mock(lines)
-        channame = 'sip/something-suffix'
-
-        result = find_participant(confd, channame)
-
-        assert_that(result, none())
-
-    def test_find_participants_when_line_has_user(self):
-        user = {
-            'uuid': 'user_uuid',
-            'tenant_uuid': 'tenant_uuid',
-            'userfield': 'user_userfield, toto',
-        }
-        lines = [{'id': 12, 'users': [user], 'extensions': []}]
-        confd = confd_mock(lines)
-        channame = 'sip/something-suffix'
-
-        result = find_participant(confd, channame)
-
-        assert_that(
-            result,
-            has_entries(
-                uuid='user_uuid',
-                tenant_uuid='tenant_uuid',
-                line_id=12,
-                tags=['user_userfield', 'toto'],
-            ),
-        )
 
 
 class TestExtractMixmonitorExtra:
